Refactored relative pathes to absolute ones
