--- src/components/MainNavigationItem/index.js
@@ -1,3 +1,0 @@
-import MainNavigationItem from './MainNavigationItem'
-
-export default MainNavigationItem
